Fix Deadlock and Reduce Memory Usage

+++ yolo/tools/data_augmentation.py
@@ -224,7 +224,7 @@ class Colorspace:
             dsaturation = np.random.uniform(1/saturation, saturation)
             dexposure   = np.random.uniform(1/exposure,   1)
 
-            hsv_image = cv2.cvtColor(image, cv2.COLOR_RGB2HSV).astype(np.float64)
+            hsv_image = cv2.cvtColor(image, cv2.COLOR_RGB2HSV).astype(np.float32)
             hsv_image[:, :, 1] *= dsaturation
             hsv_image[:, :, 2] *= dexposure
